seperate plotting directory for plotting functions
