+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/science_verification_analysis.py
@@ -16,7 +16,6 @@ from matplotlib.patches import Patch
 import matplotlib.dates as mdates
 from datetime import timezone
 from matplotlib import colors, cm
-from NuRadioReco.modules.RNO_G.channelBlockOffsetFitter import fit_block_offsets
 from config_files_sva.config_station import get_station_config, sampling_rate
 from config_files_sva.config_plotting import set_plot_style
 from read_rnog_data_nuradio import convert_events_information, read_rnog_data
@@ -30,6 +29,12 @@ from analysis_functions_sva.glitching_analysis_sva import binomtest_glitch_fract
 from config_files_sva.config_glitching import config_glitching_values
 from analysis_functions_sva.block_offsets_analysis_sva import get_block_offsets_after_removal, get_block_offsets_before_removal, plot_block_offsets_violin_before_after_comparison, block_offset_statistics
 from config_files_sva.config_block_offsets import block_offset_limits
+from plotting_functions_sva.plotting_sva_spectrum import plot_time_integrated_surface_spectra_unnormalized, plot_time_integrated_surface_spectra_normalized, plot_time_integrated_deep_spectra
+from plotting_functions_sva.plotting_sva_snr import choose_day_interval, plot_snr_against_time
+from plotting_functions_sva.plotting_sva_vrms import plot_vrms_values_against_time
+from plotting_functions_sva.plotting_sva_glitch import glitching_violin_plot, choose_bin_size, plot_glitch_q99_over_time
+from plotting_functions_sva.plotting_sva_debug import debug_plot_ratios, debug_plot_snr_distribution, debug_plot_z_score_snr, debug_plot_vrms_distribution
+
 
 #### Script directory for json files
 SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
@@ -102,267 +107,6 @@ def read_rnog_runtable(station_id: int, start_time: str, stop_time: str):
     RunTable = rt.RunTable()
     testrt = RunTable.get_table( start_time=start_time, stop_time=stop_time, stations=[station_id], run_types = ['physics'])
     return testrt
-
-def plot_time_integrated_surface_spectra_unnormalized(station_id, spec_arr, freqs, upward_channels, downward_channels, save_location, run_label, trigger_label):
-    '''Plot time-integrated surface channel spectra.'''
-    plt.figure(figsize=(10, 6))
-    for ch in upward_channels:
-        spec_mean = np.mean(spec_arr[ch, :, :], axis=0)
-        plt.plot(freqs / units.MHz, spec_mean, label=f'Ch {ch} (up)', linestyle='-')
-    for ch in downward_channels:
-        spec_mean = np.mean(spec_arr[ch, :, :], axis=0)
-        plt.plot(freqs / units.MHz, spec_mean, label=f'Ch {ch} (down)', linestyle='--')
-
-    plt.xlabel('Frequency [MHz]')
-    plt.xlim(50, 800)
-    plt.ylim(0, 5)
-    plt.ylabel('Amplitude Spectrum [V/GHz]')
-    plt.title(f'Time-Integrated Spectrum of Surface Channels  ({trigger_label} Trigger)')
-    plt.legend(loc="upper right", 
-               frameon=True,
-               fancybox=True,
-               framealpha=0.9,
-               edgecolor="black")
-    plt.grid()
-    plt.tight_layout()
-    plt.savefig(os.path.join(save_location, f"{trigger_label}_time_integrated_surface_spectra_unnormalized_{station_id}_{run_label}.pdf"))
-
-
-def plot_time_integrated_surface_spectra_normalized(station_id, norm_spec_arr, freqs, upward_channels, downward_channels, save_location, run_label):
-    '''Plot time-integrated surface channel spectra.'''
-    plt.figure(figsize=(10, 6))
-    for ch in upward_channels:
-        spec_mean = np.mean(norm_spec_arr[ch, :, :], axis=0)
-        plt.plot(freqs / units.MHz, spec_mean, label=f'Ch {ch} (up)', linestyle='-')
-    for ch in downward_channels:
-        spec_mean = np.mean(norm_spec_arr[ch, :, :], axis=0)
-        plt.plot(freqs / units.MHz, spec_mean, label=f'Ch {ch} (down)', linestyle='--')
-
-    periodiccolor2 = "mediumseagreen"
-    excesscolor = 'grey'
-    wb_color = "mediumvioletred"
-    normcolor = "steelblue"
-
-    plt.axvspan(80, 120, color=excesscolor, alpha=0.3, label="_nolegend_")
-    plt.axvline(x=0.403e3, color=wb_color, linestyle='--', linewidth=1.2, label="_nolegend_", alpha=0.7)
-    plt.axvspan(0.278e3, 0.285e3, color=periodiccolor2, alpha=0.3, label="_nolegend_")
-    plt.axvspan(0.482e3, 0.485e3, color=periodiccolor2, alpha=0.3, label="_nolegend_")
-    plt.axvspan(0.240e3, 0.272e3, color=periodiccolor2, alpha=0.3, label="_nolegend_")
-    plt.axvspan(0.360e3, 0.380e3, color=periodiccolor2, alpha=0.3, label="_nolegend_")
-    plt.axvspan(0.136e3, 0.139e3, color=periodiccolor2, alpha=0.3, label="_nolegend_")
-    plt.axvspan(0.151e3, 0.157e3, color=periodiccolor2, alpha=0.3, label="_nolegend_")
-    plt.axvspan(0.125e3, 0.127e3, color=periodiccolor2, alpha=0.3, label="_nolegend_")
-    plt.axvspan(500, 650, color=normcolor, alpha=0.3, label="_nolegend_")
-
-    plt.xlabel('Frequency [MHz]')
-    plt.xlim(50, 800)
-    plt.ylim(0, 5)
-    plt.ylabel('Amplitude Spectrum [V/GHz]')
-    plt.title('Time-Integrated Spectrum of Surface Channels (FORCE Trigger)')
-    
-    ax = plt.gca()
-    line_legend = ax.legend(
-        loc="upper right",
-        frameon=True,
-        fancybox=True,
-        framealpha=0.9,
-        edgecolor="black")
-    
-    annotation_handles = [
-        Patch(facecolor=excesscolor, alpha=0.3, label="Galactic Excess"),
-        Line2D([0], [0], color=wb_color, linestyle="--", linewidth=1.2, label="Weather Balloon"),
-        Patch(facecolor=periodiccolor2, alpha=0.3, label="Periodic Signal"),
-        Patch(facecolor=normcolor, alpha=0.3, label="Normalization Region"),]
-
-    annotation_legend = ax.legend(
-        handles=annotation_handles,
-        loc="lower right",
-        frameon=True,
-        fancybox=True,
-        framealpha=0.9,
-        edgecolor="black")
-
-    ax.add_artist(line_legend)
-
-    plt.grid()
-    plt.tight_layout()
-    plt.savefig(os.path.join(save_location, f"time_integrated_surface_spectra_normalized_force_trigger_{station_id}_{run_label}.pdf"))
-
-def plot_time_integrated_deep_spectra(station_id, spec_arr, freqs, vpol_channels, hpol_channels, save_location, run_label, trigger_label):
-    '''Plot time-integrated deep channel spectra.'''
-    plt.figure(figsize=(10, 6))
-    for ch in vpol_channels:
-        spec_mean = np.mean(spec_arr[ch, :, :], axis=0)
-        plt.plot(freqs / units.MHz, spec_mean, label=f'Ch {ch} (VPOL)', linestyle='-')
-    for ch in hpol_channels:
-        spec_mean = np.mean(spec_arr[ch, :, :], axis=0)
-        plt.plot(freqs / units.MHz, spec_mean, label=f'Ch {ch} (HPOL)', linestyle='--')
-
-    plt.xlabel('Frequency [MHz]')
-    plt.xlim(50, 800)
-    plt.ylim(0, 5)
-    plt.ylabel('Amplitude Spectrum [V/GHz]')
-    plt.title(f'Time-Integrated Spectrum of Deep Channels ({trigger_label} Trigger)')
-    plt.legend(loc="upper right", 
-               frameon=True,
-               fancybox=True,
-               framealpha=0.9,
-               edgecolor="black")
-    plt.grid()
-    plt.tight_layout()
-    plt.savefig(os.path.join(save_location, f"{trigger_label}_time_integrated_deep_spectra_unnormalized_{station_id}_{run_label}.pdf"))
-
-#### SNR analysis functions ####       
-def choose_day_interval(times):
-    times = pd.to_datetime(times, utc=True)
-    total_days = (times.max() - times.min()).days
-
-    if total_days < 10:
-        return 1
-    elif total_days < 20:
-        return 2
-    elif total_days < 40:
-        return 4
-    elif total_days < 80:
-        return 7
-    elif total_days < 150:
-        return 10
-    elif total_days < 300:
-        return 15
-    elif total_days < 600:
-        return 30
-    else:
-        return 60
-
-def plot_snr_against_time(station_id,times,snr_arr,flag,z_log,k_list,channels,save_location,run_label,nrows=12,ncols=2,day_interval=None):
-    times = pd.to_datetime(times,utc=True)
-    channels = list(channels)
-    n_channels = len(channels)
-
-    if day_interval is None:
-        day_interval = choose_day_interval(times) 
-
-    fig, axs = plt.subplots(nrows,ncols,figsize=(15,24),sharex=True)
-    axs = np.array(axs)
-
-    for idx, ch in enumerate(channels):
-        r = idx//ncols
-        c = idx%ncols
-        ax = axs[r,c]
-        good_mask = ~flag[ch]
-        ax.scatter(times[good_mask], np.log10(snr_arr[ch][good_mask]), s=8,alpha=0.25, color="gray")
-        zex = np.abs(z_log[ch]) - k_list[ch]
-        zex = np.clip(zex,0,None)
-        sc = ax.scatter(times[flag[ch]], np.log10(snr_arr[ch][flag[ch]]), s=8,c=zex[flag[ch]], cmap="Reds")
-        cax = ax.inset_axes([1.02,0.1,0.05,0.8])
-        plt.colorbar(sc,cax=cax)
-        ax.grid(alpha=0.4)
-        ax.text(0.85, 0.95, f"Ch {ch}", transform = ax.transAxes, ha = "left",va = "top", bbox = dict(boxstyle = "round, pad = 0.25", facecolor = "white", alpha = 0.8))
-    
-    for idx in range(n_channels, nrows*ncols):
-        r = idx//ncols
-        c = idx%ncols
-        axs[r, c].set_visible(False)
-
-    red = plt.cm.Reds(0.6)
-    legend_handles = [Line2D([0],[0],marker="o",color="none",markeredgecolor="gray",markerfacecolor="gray",markersize=6,label=r"$|z|\leq k$"),
-                    Line2D([0],[0],marker="o",color="none",markeredgecolor=red,markerfacecolor=red,markersize=6,label=r"$|z|>k$")]                 
-    axs[0, 0].legend(handles = legend_handles, loc = "upper left")
-
-    ticks_ax = axs[-1,0]
-    time_span = (times.max() - times.min()).total_seconds() / 86400.0  # in days
-
-    if time_span < 1:
-        # Use 2h ticks if less than 1 day
-        ticks_ax.xaxis.set_major_locator(mdates.HourLocator(interval=6))
-        ticks_ax.xaxis.set_major_formatter(mdates.DateFormatter("%m-%d\n%H:%M", tz=timezone.utc))
-    elif time_span < 3:
-        # Use 6h ticks if less than 3 days
-        ticks_ax.xaxis.set_major_locator(mdates.HourLocator(interval=12))
-        ticks_ax.xaxis.set_major_formatter(mdates.DateFormatter("%m-%d\n%H:%M", tz=timezone.utc))
-    else:
-        # Use day ticks otherwise
-        ticks_ax.xaxis.set_major_locator(mdates.DayLocator(interval = day_interval))
-        ticks_ax.xaxis.set_major_formatter(mdates.DateFormatter("%m-%d", tz = timezone.utc))
-
-    fig.autofmt_xdate()
-
-    plt.subplots_adjust(bottom = 0.11, wspace = 0.38, left=0.08)
-    fig.supxlabel("Date [UTC]", x = 0.5, y = 0.06)
-    fig.supylabel(r"$\log_{10}(\mathrm{SNR})$", x = 0.02)
-    plt.savefig(os.path.join(save_location,f"snr_against_time_{station_id}_{run_label}.pdf"))
-            
-
-#### Vrms analysis functions ####
-def plot_vrms_values_against_time(times, vrms_arr_all, channel_list, station_id, run_label, save_location, n_rows = 12, n_cols = 2, day_interval=5):
-    '''Plot Vrms distributions for different trigger types.'''
-
-    force_mask = event_info["triggerType"] == "FORCE"
-    radiant0_mask = event_info["triggerType"] == "RADIANT0"
-    radiant1_mask = event_info["triggerType"] == "RADIANT1"
-    lt_mask = event_info["triggerType"] == "LT"
-
-    trigger_masks = {"FORCE": force_mask, 
-                     "RADIANT0": radiant0_mask,
-                     "RADIANT1": radiant1_mask,
-                     "LT": lt_mask,}
-
-    trigger_colors = {"FORCE": "tab:blue", 
-                      "RADIANT0": "tab:orange",
-                      "RADIANT1": "tab:green",
-                      "LT": "tab:red",}
-
-    times = np.asarray(times)
-    vrms_arr_all = np.asarray(vrms_arr_all)
-
-    n_channels = len(channel_list)
-
-    fig, axs = plt.subplots(n_rows, n_cols, figsize=(15, 24), sharex=True, squeeze=False)
-    axs = axs.ravel()
-
-    legend_handles = {}
-
-    for idx, ch in enumerate(channel_list):
-        ax = axs[idx]
-        vrms_ch = vrms_arr_all[ch]
-
-        for trig_name, trig_mask in trigger_masks.items():
-            times_trig = times[trig_mask]
-            vrms_trig = vrms_ch[trig_mask]
-
-            scatter = ax.scatter(times_trig, vrms_trig, s=8, alpha=0.5, label=trig_name, color=trigger_colors[trig_name])
-            if trig_name not in legend_handles:
-                legend_handles[trig_name] = scatter
-                
-
-        ax.set_title(f"Channel {ch}")
-        ax.grid(alpha=0.4)
-    
-    for j in range(len(channel_list), len(axs)):
-        axs[j].set_visible(False)
-
-    ticks_ax = axs[-1]   
-    time_span = times.max() - times.min()     
-    time_span_days = time_span / np.timedelta64(1, "D")
-
-    if time_span_days < 1:
-        ticks_ax.xaxis.set_major_locator(mdates.HourLocator(interval=6))
-        ticks_ax.xaxis.set_major_formatter(mdates.DateFormatter("%m-%d\n%H:%M", tz=timezone.utc))
-    elif time_span_days < 3:
-        ticks_ax.xaxis.set_major_locator(mdates.HourLocator(interval=12))
-        ticks_ax.xaxis.set_major_formatter(mdates.DateFormatter("%m-%d\n%H:%M", tz=timezone.utc))
-    else:
-        ticks_ax.xaxis.set_major_locator(mdates.DayLocator(interval=day_interval))
-        ticks_ax.xaxis.set_major_formatter(mdates.DateFormatter("%m-%d", tz=timezone.utc))
-
-    fig.legend(handles=[legend_handles[k] for k in trigger_masks.keys()], labels=list(trigger_masks.keys()), 
-               loc="lower center", ncol=4, frameon=True, markerscale=2, bbox_to_anchor=(0.5, 0.005))
-    
-    fig.supylabel(r"$V_\mathrm{rms}$ [V]", x=0.02)
-    plt.subplots_adjust(bottom = 0.05, wspace = 0.15, left=0.1)
-    fig.supxlabel("Date [UTC]", x = 0.5, y = 0.06)
-    plt.savefig(os.path.join(save_location, f"vrms_against_time_{station_id}_{run_label}.pdf"))
-    plt.close(fig)
 
 #### Trigger analysis (adapted from the plot_trigger() function from analyze_run.py) ####
 def compute_radiant_thresholds(event_info, down_channels, up_channels):
@@ -455,186 +199,6 @@ def plot_trigger_rate_with_thresholds(station_id, event_info, down_channels, up_
 
     return fig, ax_rate, ax_thr
 
-#### Glitching analysis functions ####
-def glitching_violin_plot(glitch_arr, channel_list, station_id, run_label, save_location):
-    data = [glitch_arr[ch] for ch in channel_list]
-    means = np.array([np.mean(glitch_arr[ch]) for ch in channel_list])
-
-    fig, ax = plt.subplots(figsize=(12, 10))
-
-    parts = ax.violinplot(data, positions=channel_list, showextrema=True, showmedians=True, vert=False, side="high", widths=1.8,)
-
-    norm = colors.Normalize(vmin=np.min(means), vmax=np.max(means), clip=False)
-    sm = cm.ScalarMappable(norm=norm, cmap="Blues")
-
-    for idx, pc in enumerate(parts["bodies"]):
-        pc.set_facecolor(sm.to_rgba(means[idx]))
-        pc.set_alpha(0.5)
-        pc.set_edgecolor("k")
-
-    parts["cmins"].set_linewidth(0.2)
-    parts["cmaxes"].set_linewidth(0.2)
-    parts["cbars"].set_linewidth(0.5)
-    parts["cmins"].set_color("k")
-    parts["cmaxes"].set_color("k")
-    parts["cbars"].set_color("k")
-    parts["cmedians"].set_color("k")
-    parts["cmedians"].set_linewidth(1)
-
-    cb = plt.colorbar(sm, ax=ax, pad=0.02)
-    cb.set_label("Mean test statistics")
-
-    ax.set_xlabel("Glitching test statistics")
-    ax.set_ylabel("Channel")
-    ax.set_yticks(channel_list)
-    ax.grid(True, alpha=0.3)
-    ax.set_ylim(min(channel_list) - 1, max(channel_list) + 1)
-
-    plt.tight_layout()
-    plt.savefig(os.path.join(save_location, f"glitching_violin_plot_{station_id}_{run_label}.pdf"))
-    plt.close(fig)
-
-def choose_bin_size(times):
-    total_hours = (times.max() - times.min()).total_seconds() / 3600.0
-    if total_hours < 12:
-        return "30min"
-    elif total_hours < 24:
-        return "1h"
-    elif total_hours < 3 * 24:
-        return "2h"
-    elif total_hours < 7 * 24:
-        return "6h"
-    elif total_hours < 30 * 24:
-        return "12h"
-    else:
-        return "24h"
-
-def plot_glitch_q99_over_time(times, glitch_arr, channels, station_id, run_label, save_location):
-    times = pd.to_datetime(times)
-    df = pd.DataFrame(glitch_arr.T, index=times, columns=channels)
-
-    bin_rule = choose_bin_size(times)
-    q99 = df.resample(bin_rule).quantile(0.99)
-
-    fig, ax = plt.subplots(figsize=(12, 10))
-
-    for ch in channels:
-        ax.plot(q99.index, q99[ch], marker=".", linestyle="-", label=f"ch {ch}")
-
-    ax.set_xlabel("Date [UTC]")
-    ax.set_ylabel(f"99% quantile glitching ts ({bin_rule} bins)")
-    ax.grid(True, alpha=0.3)
-    ax.legend(ncol=3, frameon=True, framealpha=0.9, edgecolor="black")
-
-    plt.tight_layout()
-    plt.savefig(os.path.join(save_location, f"glitch_q99_{station_id}_{run_label}.pdf"))
-    plt.close(fig)
-
-#### Debug plots ####
-def debug_plot_ratios(ratio_arr_dict, channels_order, save_location, station_id, run_label, bins=30):
-    n_bands = len(ratio_arr_dict)
-    
-    fig, axes = plt.subplots(1, n_bands, figsize=(5*n_bands, 5), sharey=True)
-    
-    if n_bands == 1:
-        axes = [axes]
-
-    for ax, (band_name, ratio_list) in zip(axes, ratio_arr_dict.items()):
-        for ch, r in zip(channels_order, ratio_list):
-            r = np.asarray(r)
-            ax.hist(np.log10(r), bins=bins, histtype="step", linewidth=1.3, label=f"Ch {ch}", alpha=0.8)
-
-        ax.set_title(band_name)
-        ax.set_xlabel("log10(R)")
-        ax.grid(True, alpha=0.3)
-
-        ax.set_title(f"{band_name} (FORCE Trigger)")
-        ax.set_xlabel("log10(R)")
-        ax.grid(True, alpha=0.3)
-
-    axes[0].set_ylabel("Counts")
-
-    handles, labels = axes[0].get_legend_handles_labels()
-    fig.legend(handles, labels, loc="upper right", ncol=8, frameon=True)
-    plt.tight_layout()
-    fig.savefig(os.path.join(save_location,f"debug_amplitude_ratios_force_trigger_{station_id}_{run_label}.pdf",))
-    plt.close(fig)
-
-def debug_plot_snr_distribution(log_snr_arr, channel_list, save_location, station_id, run_label, bins=30):
-    fig, ax = plt.subplots(figsize=(10, 6))
-    for ch in channel_list:
-        log_snr_ch = log_snr_arr[ch]
-        ax.hist(log_snr_ch, bins=bins, histtype="step", linewidth=1.3, label=f"Ch {ch}", alpha=0.8)
-
-    ax.set_title("Log10 SNR Distribution (FORCE Trigger)")
-    ax.set_xlabel("log10(SNR)")
-    ax.set_ylabel("Counts")
-    ax.grid(True, alpha=0.3)
-
-    handles, labels = ax.get_legend_handles_labels()
-    fig.legend(handles, labels, loc="upper right", ncol=8, frameon=True)
-    plt.tight_layout()
-    fig.savefig(os.path.join(save_location,f"debug_snr_distribution_force_trigger_{station_id}_{run_label}.pdf",))
-    plt.close(fig)
-
-def debug_plot_z_score_snr(z_score_arr, channel_list, save_location, station_id, run_label, bins=30):
-    fig, ax = plt.subplots(figsize=(10, 6))
-    for ch in channel_list:
-        z_score_ch = z_score_arr[ch]
-        ax.hist(z_score_ch, bins=bins, histtype="step", linewidth=1.3, label=f"Ch {ch}", alpha=0.8)
-
-    ax.set_title("Z-Score SNR Distribution (FORCE Trigger)")
-    ax.set_xlabel("Z-Score(SNR)")
-    ax.set_ylabel("Counts")
-    ax.grid(True, alpha=0.3)
-
-    handles, labels = ax.get_legend_handles_labels()
-    fig.legend(handles, labels, loc="upper right", ncol=8, frameon=True)
-    plt.tight_layout()
-    fig.savefig(os.path.join(save_location,f"debug_z_score_snr_force_trigger_{station_id}_{run_label}.pdf",))
-    plt.close(fig)
-
-def debug_plot_vrms_distribution(vrms_arr, modality_dict, channel_list, station_id, run_label, trigger_label, save_location, n_rows=12, n_cols=2):
-    fig, axes = plt.subplots(n_rows, n_cols, figsize=(16, 36))
-    axes = axes.flatten()
-
-    for idx, ch in enumerate(channel_list):
-        ax = axes[idx]
-        info = modality_dict[ch]
-        vrms = vrms_arr[ch]
-        vrms = vrms[np.isfinite(vrms)] 
-        vrms_grid = info["vrms_grid"]
-        kde_values = info["kde_values"]
-        peaks = info["peaks"]
-        n_peaks = info["n_peaks"]
-
-        if n_peaks == 0:
-            modality = "flat/noisy"
-        elif n_peaks == 1:
-            modality = "unimodal"
-        elif n_peaks == 2:
-            modality = "bimodal"
-        else:
-            modality = f"multimodal ({n_peaks})"
-
-        # histogram
-        ax.hist(vrms, bins=30, density=True, alpha=0.3, color="gray")
-        # kde curve
-        ax.plot(vrms_grid, kde_values, color="blue", lw=1.5)
-        # peaks
-        if len(peaks) > 0:
-            ax.plot(vrms_grid[peaks], kde_values[peaks], "ro", markersize=5)
-
-        ax.set_title(f"Ch {ch}: {modality}")
-        ax.set_xlabel("Vrms Values [V]")
-        ax.set_ylabel("KDE Density")
-
-    for i in range(len(channel_list), n_rows * n_cols):
-        axes[i].axis("off")
-
-    plt.tight_layout()
-    plt.savefig(os.path.join(save_location,f"debug_vrms_hist_kde_density_peaks_{station_id}_{run_label}_{trigger_label}.pdf",))
-
 def channel_health(row):
     severity = {"OK": 0, "!!": 1, "X": 2}
     vals = [severity[v] for v in row if v in severity]
@@ -642,7 +206,6 @@ def channel_health(row):
         return "-"
     inv = {0: "OK", 1: "!!", 2: "X"}
     return inv[max(vals)]
-
 
 def create_result_csv_file(station_id, run_label, n_events_force, surface_channels, downward_channels, upward_channels, all_channels, validation_results, glitch_info, modality_dict_force, modality_dict_lt, 
                            modality_dict_radiant0, modality_dict_radiant1, outlier_details, save_location):
@@ -1103,7 +666,7 @@ if __name__ == "__main__":
     if len(vrms_arr_lt[1]) < 100:
             logger.warning(f"LT trigger has less than 100 valid Vrms entries ({len(vrms_arr_lt)}). Results for the Vrms statistics may be unreliable.")
     modality_lt, tail_label_lt = report_vrms_characteristics(modality_dict_lt, tail_dict_lt, all_channels, report_config=report_vrms_function_parameters)
-    plot_vrms_values_against_time(times, vrms_arr, all_channels, station_id, run_label, save_location, n_rows=12, n_cols=2, day_interval=day_interval)
+    plot_vrms_values_against_time(event_info, times, vrms_arr, all_channels, station_id, run_label, save_location, n_rows=12, n_cols=2, day_interval=day_interval)
 
     # Write detailed Vrms modality results to text files for each trigger type
     write_vrms_modality_results(modality_force, tail_label_force, trigger_label="FORCE", station_id=station_id, run_label=run_label)
